Fix: Take u32 as an offset

+++ src/builder.rs
@@ -295,15 +295,15 @@ impl<'a> Builder<'a> {
         Ok(self.inst_builder.build_array_alloca(t.to_cl()?, self.number_constant(size)?, "array_alloc"))
     }
 
-    pub fn store(&mut self, v: Value, addr: values::PointerValue, offset: i32) -> Result<(), Error> {
-        let ptr = self.inst_builder.build_in_bounds_gep(addr, &[types::IntType::i32_type().const_int(0, false), types::IntType::i32_type().const_int(offset, false)], "store");
+    pub fn store(&mut self, v: Value, addr: values::PointerValue, offset: u32) -> Result<(), Error> {
+        let ptr = self.inst_builder.build_in_bounds_gep(addr, &[types::IntType::i32_type().const_int(0, false), types::IntType::i32_type().const_int(offset as u64, false)], "store");
         let cl = self.to_cl(v)?;
         self.inst_builder.build_store(ptr, cl);
         Ok(())
     }
 
-    pub fn load(&mut self, t: Type, addr: values::PointerValue, offset: i32) -> Result<Value, Error> {
-        let ptr = self.inst_builder.build_in_bounds_gep(addr, &[types::IntType::i32_type().const_int(0, false), types::IntType::i32_type().const_int(offset, false)], "store");
+    pub fn load(&mut self, t: Type, addr: values::PointerValue, offset: u32) -> Result<Value, Error> {
+        let ptr = self.inst_builder.build_in_bounds_gep(addr, &[types::IntType::i32_type().const_int(0, false), types::IntType::i32_type().const_int(offset as u64, false)], "store");
         let data = ValueData::from_cl(self.inst_builder.build_load(ptr, "load"), t.cl_type()?)?;
         Ok(self.value_store.new_value(data))
     }
